% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Function to run the analysis}
\usage{
analyze(
  n,
  n_captures = 6,
  n_strata,
  p_captures,
  p_strata,
  suppress,
  method = "poisson",
  formula.selection = "stepwise",
  opts.stepwise = list(direction = "forward", threshold = 0.5, verbose = FALSE)
)
}
\arguments{
\item{n}{int: total number entering the simulation}

\item{suppress}{bool: mimic required data suppression (N_ID in 1:10 becomes obfuscated)}

\item{formula.selection}{string: selection for formula decision - either "aic", "corr", or "stepwise"}

\item{opts.stepwise}{list: list containing 'direction' of 'forward' 'backward' or 'both', and 'threshold': p value threshold for stepwise selection}

\item{correlate}{bool: attempts to correlate captures (default is analagous to MA PHD data)}

\item{groups}{string: by-group to estimate
#' @param method string: selection for spatial capture recapture model - either "poisson" or "negbin"}
}
\description{
Function to run the analysis
}
