% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossfold.R
\name{build_folds}
\alias{build_folds}
\title{Build cross-validation folds and run estimation for each fold.}
\usage{
build_folds(fold, data, nfolds, permutation)
}
\arguments{
\item{fold}{The current fold number.}

\item{data}{The dataset to be split into folds.}

\item{nfolds}{The total number of folds.}

\item{permutation}{A permutation of row indices for creating folds.}
}
\value{
A list containing the training set, test set, and the subset of
indices for the current fold.
}
\description{
Build cross-validation folds and run estimation for each fold.
}
