% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_control.R
\name{build_individual_records}
\alias{build_individual_records}
\title{Build Individual Simulation Records. This builds all the individual records
for a simulated dataset. Each record produced by this function represents
one individual and contains the capture indicators and simulated numerical
covariate indicators. This function is traditionally wrapped by
\code{create_data} but
exposed for users who want more granular control over building the simulation
data. Importantly, the capture_probs and covariate_specs are not validated so
errors may not be as descriptive.}
\usage{
build_individual_records(n_individuals, capture_probs, covariate_specs)
}
\arguments{
\item{n_individuals}{int: number of participants to simulate}

\item{capture_probs}{numeric vector: capture probabilities of length equal to
the number of capture columns}

\item{covariate_specs}{data.frame: covariate specification table with one row
per covariate and columns \code{distribution}, \code{p1}, \code{p2}, and
\code{dtype}.}
}
\value{
a tibble where each row is an individual record
}
\description{
Build Individual Simulation Records. This builds all the individual records
for a simulated dataset. Each record produced by this function represents
one individual and contains the capture indicators and simulated numerical
covariate indicators. This function is traditionally wrapped by
\code{create_data} but
exposed for users who want more granular control over building the simulation
data. Importantly, the capture_probs and covariate_specs are not validated so
errors may not be as descriptive.
}
\keyword{internal}
