% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_listpairs}
\alias{build_listpairs}
\title{Build all unique pairs of capture lists.}
\usage{
build_listpairs(n_lists)
}
\arguments{
\item{n_lists}{The number of capture lists.}
}
\value{
A character vector containing all unique pairs of capture lists in
the format "j,k".
}
\description{
Build all unique pairs of capture lists.
}
