% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_control.R
\name{captures_from_covariates}
\alias{captures_from_covariates}
\title{Sample Captures Based on Covariates. This function generates a data frame of
simulated capture data for a specified number of individuals based on the
provided covariate data. Each row represents an individual, and each column
represents a capture event. The values in the data frame are binary (0 or 1),
indicating whether an individual was captured (1) or not (0) for each
capture event. The capture probabilities are determined by the covariate
values, specifically the "sex" column, where the probability of capture is
higher for one sex compared to the other.}
\usage{
captures_from_covariates(data_table, n_cols, cov_func)
}
\arguments{
\item{data_table}{data.frame: A data frame containing covariate data for each
individual. The data frame must contain a "sex" column with binary values (0
or 1) indicating the sex of each individual.}

\item{n_cols}{int: The number of capture columns to generate. Each column
represents a capture event.}

\item{covariate_conditions}{function: function that takes \code{data_table}
and returns a numeric probability vector \code{capture_prob} in [0, 1].}
}
\value{
A data frame where each row represents an individual and each column
represents a capture event. The values are binary (0 or 1), indicating
whether an individual was captured (1) or not (0) for each capture event,
based on the covariate values in the input data frame.
}
\description{
Sample Captures Based on Covariates. This function generates a data frame of
simulated capture data for a specified number of individuals based on the
provided covariate data. Each row represents an individual, and each column
represents a capture event. The values in the data frame are binary (0 or 1),
indicating whether an individual was captured (1) or not (0) for each
capture event. The capture probabilities are determined by the covariate
values, specifically the "sex" column, where the probability of capture is
higher for one sex compared to the other.
}
