% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Function to compare between simulation results and ground truth}
\usage{
compare(sim, groundTruth)
}
\arguments{
\item{sim}{list of simulations done by \code{simulate}}

\item{groundTruth}{list of ground truths by \code{extract.groundTruth}}
}
\description{
Function to compare between simulation results and ground truth
}
\keyword{internal}
