% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_capture.R
\name{conditional_capture}
\alias{conditional_capture}
\title{Generate the conditional capture probability function.}
\usage{
conditional_capture(q_1, q_2, q_12)
}
\arguments{
\item{q_1}{The nuisance function q_1.}

\item{q_2}{The nuisance function q_2.}

\item{q_12}{The nuisance function q_12.}
}
\value{
A function that computes the conditional capture probability given
covariate data.
}
\description{
Generate the conditional capture probability function.
}
\references{
Das, Manjari and Kennedy, Edward H. and Jewell, Nicholas P.
(2024). "Doubly Robust Capture-Recapture Methods for Estimating Population
Size". Journal of the American Statistical Association, 119(546), 1309-1321.
}
