% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_specs.R
\name{continuous_heterogeneity_spec}
\alias{continuous_heterogeneity_spec}
\title{Create a continuous latent heterogeneity specification}
\usage{
continuous_heterogeneity_spec(sd)
}
\arguments{
\item{sd}{Standard deviation of the shared latent propensity.}
}
\value{
An S3 object with class `continuous_heterogeneity_spec`.
}
\description{
Create a continuous latent heterogeneity specification
}
