% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_specs.R
\name{covariate_capture_spec}
\alias{covariate_capture_spec}
\title{Create a covariate-dependent capture specification}
\usage{
covariate_capture_spec(covariate_ranges, alpha, beta)
}
\arguments{
\item{covariate_ranges}{Data frame describing generated covariates.}

\item{alpha}{List-specific logistic intercepts.}

\item{beta}{List-by-covariate logistic coefficient matrix.}
}
\value{
An S3 object with class `covariate_capture_spec`.
}
\description{
Create a covariate-dependent capture specification
}
