% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcsim-package.R
\docType{package}
\name{crcsim-package}
\alias{crcsim}
\alias{crcsim-package}
\title{crcsim: Simulating Data Capture Recapture}
\description{
Simulating Data Capture Recapture (crcsim) focuses on simulating data for capture-recapture (crc) analyses. The package includes functions to generate crc data, extract the unknown values from the generated data, fit multiple different log-linear model configurations to the data, and compare the estimated intercept value to the ground truth. The package is designed to be flexible and can be used for a variety of applications, including epidemiology, ecology, and social sciences. The package can be used with either generated or user provided data. The generated data can be customized to include different numbers of captures, different capture probabilities, and different levels of heterogeneity. The user-provided data can be in the form of a contingency table or a data frame with binary indicators for each capture.
}
\keyword{internal}
