% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create.capture}
\alias{create.capture}
\title{A function to generate binomial flips along the probability specified}
\usage{
create.capture(n, prob)
}
\arguments{
\item{n}{number of 'captures' to simulate}

\item{prob}{list of probabilities respective to each \code{n}}
}
\description{
A function to generate binomial flips along the probability specified
}
\keyword{internal}
