% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create.data}
\alias{create.data}
\title{Data generation tool}
\usage{
create.data(n_individuals, p_captures, p_strata)
}
\arguments{
\item{n_individuals}{int: number of participants to simulate}

\item{p_captures}{list: list of capture probabilities to emulate}

\item{p_strata}{list: list of extra 'stratification' probabilities to emulate}
}
\value{
a data.table
}
\description{
Data generation tool
}
\keyword{internal}
