% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create.strata}
\alias{create.strata}
\title{A function to sample from 1:n along probability \code{prob}}
\usage{
create.strata(n, prob)
}
\arguments{
\item{n}{number of 'strata' to simulate}

\item{prob}{probabilities of n - expects summation to 1}
}
\description{
A function to sample from 1:n along probability \code{prob}
}
\keyword{internal}
