% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubly_robust.R
\name{doubly_robust_estimator}
\alias{doubly_robust_estimator}
\title{Compute the doubly robust estimator for capture-recapture data.}
\usage{
doubly_robust_estimator(q_1, q_2, q_12, phi)
}
\arguments{
\item{q_1}{The nuisance function q_1.}

\item{q_2}{The nuisance function q_2.}

\item{q_12}{The nuisance function q_12.}

\item{phi}{The auxiliary function phi.}
}
\value{
The doubly robust estimate of the capture-recapture parameter.
}
\description{
Compute the doubly robust estimator for capture-recapture data.
}
