% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{estimate.age}
\alias{estimate.age}
\title{Draws age from a sample of bins}
\usage{
estimate.age(
  age = 1:10,
  prob = c(0.00295, 0.0271, 0.273, 0.274, 0.194, 0.154, 0.0512, 0.0129, 0.00568,
    0.000681)
)
}
\arguments{
\item{age}{list: age bins to sample from}

\item{prob}{list: probabilities relative to \code{age}}

\item{seed}{int: seeding}
}
\description{
Draws age from a sample of bins
}
\keyword{internal}
