% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{estimate.captures}
\alias{estimate.captures}
\title{Draws captures as a series of binomial flips}
\usage{
estimate.captures(
  capture = c("APCD", "BSAS", "Casemix", "Death", "Matris", "PMP"),
  prob = c(0.4887, 0.1467, 0.096, 0.0108, 0.0527, 0.205),
  correlate = FALSE,
  corr.matrix = matrix(c(1, 0.03464933, -0.09139326, -0.06406634, -0.10679959,
    -0.1156107, 0.03464933, 1, 0.210205, -0.06738847, 0.33085861, 0.05487056,
    -0.09139326, 0.210205, 1, -0.08669687, 0.09550053, -0.00245505, -0.06406634,
    -0.06738847, -0.08669687, 1, -0.08148882, -0.04910034, -0.10679959, 0.33085861,
    0.09550053, -0.08148882, 1, 0.11940396, -0.1156107, 0.05487056, -0.00245505,
    -0.04910034, 0.11940396, 1), ncol = 6, byrow = TRUE),
  seed
)
}
\arguments{
\item{capture}{list: captures to sample from}

\item{prob}{list: binomial probabilities relative to \code{capture}}

\item{correlate}{bool: attempts correlation between captures}

\item{corr.matrix}{matrix: correlation between \code{capture}}

\item{seed}{int: seeding}
}
\description{
Draws captures as a series of binomial flips
}
\keyword{internal}
