% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{estimate.race}
\alias{estimate.race}
\title{Draws race from a sample}
\usage{
estimate.race(
  race = c("White", "Black", "Asian/PI", "Hispanic", "Other"),
  prob = c(0.777, 0.0509, 0.00629, 0.0942, 0.0109)
)
}
\arguments{
\item{race}{list: races to sample from}

\item{prob}{list: sampling probabilities relative to \code{race}}

\item{seed}{int: seeding}
}
\description{
Draws race from a sample
}
\keyword{internal}
