% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{estimate.sex}
\alias{estimate.sex}
\title{Draws sex from a sample}
\usage{
estimate.sex(sex = c("Male", "Female"), prob = c(0.583, 0.386))
}
\arguments{
\item{sex}{list: sexes to sample from}

\item{prob}{list: sampling probabilities relative to \code{sex}}

\item{seed}{int: seeding}
}
\description{
Draws sex from a sample
}
\keyword{internal}
