% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_probability.R
\name{estimate_no_covariates}
\alias{estimate_no_covariates}
\title{Estimate population size without covariates.}
\usage{
estimate_no_covariates(data, n_lists)
}
\arguments{
\item{data}{The dataset containing capture columns.}

\item{n_lists}{The number of capture lists.}
}
\value{
A list containing the estimation results and the total population
size.
}
\description{
Estimate population size without covariates.
}
