% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_probability.R
\name{estimate_psi}
\alias{estimate_psi}
\title{Estimate the capture probability and its variance using the specified method.}
\usage{
estimate_psi(method, nuisance_functions, y_j, y_k, margin, n_lists, n)
}
\arguments{
\item{method}{A character string specifying the estimation method ("tmle" or
"doubly_robust").}

\item{nuisance_functions}{A list containing the nuisance functions q_1, q_2,
and q_12.}

\item{y_j}{A numeric vector representing the observations for capture list j.}

\item{y_k}{A numeric vector representing the observations for capture list k.}

\item{margin}{A numeric value specifying the margin parameter.}

\item{n_lists}{An integer specifying the number of capture lists.}

\item{n}{An integer specifying the sample size.}
}
\value{
A list containing the estimated capture probability and its variance.
}
\description{
Estimate the capture probability and its variance using the specified method.
}
\keyword{internal}
