% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{EstimatorOptions}
\alias{EstimatorOptions}
\title{Plugin Options Class}
\description{
This class defines the options for the plugin estimator.
}
\section{Super class}{
\code{\link[crcsim:Options]{Options}} -> \code{EstimatorOptions}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{nuisance_function}}{Character scalar identifying the function
to use to estimate nuisance parameters.}

    \item{\code{nfolds}}{Integer scalar giving the number of cross-validation
folds.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-EstimatorOptions-initialize}{\code{EstimatorOptions$new()}}
    \item \href{#method-EstimatorOptions-clone}{\code{EstimatorOptions$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EstimatorOptions-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-EstimatorOptions-initialize}{}}}
\subsection{\code{EstimatorOptions$new()}}{
  Create a new \code{EstimatorOptions} instance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EstimatorOptions$new(
  method,
  capture_columns,
  threshold,
  nuisance_function,
  nfolds
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{method}}{Character scalar naming the estimation method.}
      \item{\code{capture_columns}}{Character vector naming the binary capture indicator
columns.}
      \item{\code{threshold}}{Numeric scalar giving the probability margin.}
      \item{\code{nuisance_function}}{Character scalar identifying the nuisance model.}
      \item{\code{nfolds}}{Integer scalar giving the number of cross-validation folds.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EstimatorOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EstimatorOptions-clone}{}}}
\subsection{\code{EstimatorOptions$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EstimatorOptions$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
