% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract.groundTruth}
\alias{extract.groundTruth}
\title{Function to extract the ground truth from data created through the \code{create.data} method}
\usage{
extract.groundTruth(
  DT,
  capture = c("APCD", "BSAS", "Casemix", "Death", "Matris", "PMP"),
  group
)
}
\arguments{
\item{DT}{data.table from \code{create.data}}

\item{capture}{list: strings of captures}

\item{group}{list: strings of groups}
}
\value{
list the same length of \code{group}
}
\description{
Function to extract the ground truth from data created through the \code{create.data} method
}
