% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nuisance.R
\name{fit_glm_tmle}
\alias{fit_glm_tmle}
\title{Fit a GLM for TMLE updates.}
\usage{
fit_glm_tmle(response, offset, ratio)
}
\arguments{
\item{response}{The response variable.}

\item{offset}{The offset for the GLM.}

\item{ratio}{The ratio used as the predictor in the GLM.}
}
\value{
A list containing the fitted values and the maximum absolute
coefficient value as the error.
}
\description{
Fit a GLM for TMLE updates.
}
\keyword{internal}
