% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglinear.R
\name{fit_loglinear_model}
\alias{fit_loglinear_model}
\title{Fit a log-linear model to the provided data using the specified formula.
Generally used as a helper function within the CRC function to fit models
for different formula specifications. The user can specify the model family
to use, either "poisson" or "negbin".}
\usage{
fit_loglinear_model(
  model_data,
  formula_object,
  model_family = c("poisson", "negbin")
)
}
\arguments{
\item{formula_object}{a formula object specifying the log-linear model to
fit.}

\item{model_family}{a string specifying the model family to use, either
"poisson" or "negbin". Default is "poisson".}

\item{data}{a data frame containing the observed capture histories and
frequency column.}
}
\value{
a fitted model object resulting from the log-linear model fit.
}
\description{
Fit a log-linear model to the provided data using the specified formula.
Generally used as a helper function within the CRC function to fit models
for different formula specifications. The user can specify the model family
to use, either "poisson" or "negbin".
}
