% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{formula_list}
\alias{formula_list}
\title{Generates All Possible Combination of Interaction Terms
Function that returns a list of every possible combination of
interaction terms}
\usage{
formula_list(y, x)
}
\arguments{
\item{y}{Column: Column on the estimation side of the equation.}

\item{x}{List: Columns to return every combination of}
}
\description{
Generates All Possible Combination of Interaction Terms
Function that returns a list of every possible combination of
interaction terms
}
