% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_specs.R
\name{homogeneous_capture_spec}
\alias{homogeneous_capture_spec}
\title{Create a homogeneous capture specification}
\usage{
homogeneous_capture_spec(p_captures = NULL)
}
\arguments{
\item{p_captures}{Optional vector of list-specific capture probabilities.}
}
\value{
An S3 object with class `homogeneous_capture_spec`.
}
\description{
Create a homogeneous capture specification
}
