% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{influence}
\alias{influence}
\title{Generate the influence function for the given data and nuisance functions.}
\usage{
influence(list_1, list_2, q_1, q_2, q_12)
}
\arguments{
\item{list_1}{The dataset corresponding to Y_1.}

\item{list_2}{The dataset corresponding to Y_2.}

\item{q_1}{The nuisance function q_1.}

\item{q_2}{The nuisance function q_2.}

\item{q_12}{The nuisance function q_12.}

\item{covariate_data}{The covariate data for the observations.}

\item{psi}{The nuisance function psi.}
}
\value{
The influence function evaluated for the given data and nuisance
functions.
}
\description{
Generate the influence function for the given data and nuisance functions.
}
\references{
Das, Manjari and Kennedy, Edward H. and Jewell, Nicholas P.
(2024). "Doubly Robust Capture-Recapture Methods for Estimating Population
Size". Journal of the American Statistical Association, 119(546), 1309-1321.
}
