% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_control.R
\name{is_column_binary}
\alias{is_column_binary}
\title{A function used to check if a column in a data frame is binary
(i.e., contains only 0s and 1s).}
\usage{
is_column_binary(column)
}
\arguments{
\item{column}{a vector representing a column in a data frame}
}
\value{
a boolean indicating whether the column is binary
}
\description{
A function used to check if a column in a data frame is binary
(i.e., contains only 0s and 1s).
}
\keyword{internal}
