% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_control.R
\name{is_probability_scalar}
\alias{is_probability_scalar}
\title{Is Probability Scalar}
\usage{
is_probability_scalar(value)
}
\arguments{
\item{value}{a value to check}
}
\value{
a boolean indicating whether the value is a valid probability scalar
}
\description{
A function used to check if a value is a valid probability
scalar (i.e., numeric, length 1, and between 0 and 1). This is used in the
validation of capture and category probabilities for data generation when
the user specifies a single probability to be applied across all captures or
categories.
}
\keyword{internal}
