% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_control.R
\name{is_probability_vector}
\alias{is_probability_vector}
\title{Is Probability Vector}
\usage{
is_probability_vector(vec)
}
\arguments{
\item{vec}{a vector to check}
}
\value{
a boolean indicating whether the vector is a valid probability
vector
}
\description{
A function used to check if a vector is a valid probability
vector (i.e., numeric and all values between 0 and 1). This is used in the
validation of capture and category probabilities for data generation.
}
\keyword{internal}
