% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_number}
\alias{is_valid_number}
\title{Validate if a value is a numeric scalar.}
\usage{
is_valid_number(x)
}
\arguments{
\item{x}{The value to check.}
}
\value{
TRUE if the value is not a numeric scalar, FALSE otherwise.
}
\description{
Validate if a value is a numeric scalar.
}
