% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_specs.R
\name{latent_class_spec}
\alias{latent_class_spec}
\title{Create a latent-class heterogeneity specification}
\usage{
latent_class_spec(probabilities, capture_probabilities)
}
\arguments{
\item{probabilities}{Latent-class assignment probabilities.}

\item{capture_probabilities}{Class-by-capture probability matrix.}
}
\value{
An S3 object with class `latent_class_spec`.
}
\description{
Create a latent-class heterogeneity specification
}
