% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_specs.R
\name{pairwise_dependence_spec}
\alias{pairwise_dependence_spec}
\title{Create a pairwise dependence specification}
\usage{
pairwise_dependence_spec(pair, strength = 0)
}
\arguments{
\item{pair}{Ordered pair of dependent capture-list indices.}

\item{strength}{Conditional-logit dependence strength.}
}
\value{
An S3 object with class `pairwise_dependence_spec`.
}
\description{
Create a pairwise dependence specification
}
