% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{plugin_estimator}
\alias{plugin_estimator}
\title{Compute the plugin estimator for the capture probability.}
\usage{
plugin_estimator(q_1, q_2, q_12)
}
\arguments{
\item{q_1}{The nuisance function q_1.}

\item{q_2}{The nuisance function q_2.}

\item{q_12}{The nuisance function q_12.}
}
\value{
The plugin estimate of the capture probability.
}
\description{
Compute the plugin estimator for the capture probability.
}
\references{
Das, Manjari and Kennedy, Edward H. and Jewell, Nicholas P.
(2024). "Doubly Robust Capture-Recapture Methods for Estimating Population
Size". Journal of the American Statistical Association, 119(546), 1309-1321.
}
