% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_control.R
\name{relocate_covariates}
\alias{relocate_covariates}
\title{Relocate Covariates}
\usage{
relocate_covariates(data, covariate_cols)
}
\arguments{
\item{data}{A data frame containing the covariate columns to be relocated}

\item{covariate_cols}{A character vector of covariate column names to be
relocated.}
}
\value{
A data frame with the specified covariate columns relocated to the
end.
}
\description{
This function relocates specified covariate columns in a data
frame to the end of the data frame. It takes a data frame and a vector of
covariate column names as input and returns the modified data frame with the
covariate columns relocated to the end.
}
