% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_level_estimation.R
\name{row_level_estimation}
\alias{row_level_estimation}
\title{CRC with Plugin Estimators}
\usage{
row_level_estimation(data, opts, seed = NULL, diagnostics = "quiet")
}
\arguments{
\item{data}{a data frame containing the capture data. All columns must be
numeric type.}

\item{opts}{a \code{EstimatorOptions} object specifying the plugin method and
parameters for the TMLE estimation.}
}
\value{
a list containing the estimate of population size and
confidence intervals.
}
\description{
This function forms the basis for the CRC estimation process on
row level data. This usually happens when there are covariates for each
observation that we can use during the estimation process. The function
takes in a data frame and an \code{EstimatorOptions} object specifying the
plugin estimator to use and its parameters. The function first checks that
the provided options are valid and that the data is not a frequency table.
It then identifies the capture indicator columns, either from the options or
by inferring them from the data. Next, it generates cross-validation folds
and computes capture probabilities using the specified nuisance function and
plugin estimator. Finally, it calculates and returns the TMLE estimate of
population size along with confidence intervals. Supported plugin estimators
include "logit", "ranger", and "gam".
}
