% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_parallel_crc.R
\name{run_crc}
\alias{run_crc}
\title{Run CRC Models}
\usage{
run_crc(
  n_individuals = 3e+05,
  p_strata = 1,
  p_captures = c(0.45, 0.3, 0.2),
  suppress = 10,
  mode = c("single", "bootstrap"),
  n_bootstraps = 100,
  methods = c("poisson", "negbin"),
  formula_selection = "stepwise",
  opts_stepwise = default_stepwise_config(),
  opts_tmle = default_tmle_config(),
  seed = 2024
)
}
\arguments{
\item{n_individuals}{int: number of individuals sampled per run}

\item{p_strata}{list: list of strata probabilities - for nonstratified use 1}

\item{p_captures}{list: list of capture probabilities}

\item{suppress}{int: suppression cap used during simulation}

\item{mode}{string: either "single" or "bootstrap"}

\item{n_bootstraps}{int: number of bootstrap runs when mode is "bootstrap"}

\item{methods}{character vector: one or both of "poisson" and "negbin"}

\item{formula_selection}{string: formula selection approach}

\item{opts_tmle}{list: TMLE options where \code{list_pair} is required when
formula_selection is \code{"tmle"}}

\item{seed}{int: starting seed}

\item{config}{list: stepwise-selection configurations}
}
\description{
Wrapper entrypoint for package consumers. Supports either one simulation run
or bootstrap resampling across many runs.
}
