% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCRCsim.R
\name{runCRC}
\alias{runCRC}
\title{Function to run CRC sim - returns a list of data.tables resulting from \code{analyze}}
\usage{
runCRC(nboot, ncores, seed = 2024)
}
\arguments{
\item{nboot}{int: number of bootstraps}

\item{ncores}{int: number of cores to use}

\item{seed}{int: starting seed}
}
\description{
Function to run CRC sim - returns a list of data.tables resulting from \code{analyze}
}
