% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_control.R
\name{sample_captures}
\alias{sample_captures}
\title{Sample Captures for Individuals. This function generates a data frame of
simulated capture data for a specified number of individuals and capture
probabilities. Each row represents an individual, and each column represents
a capture event. The values in the data frame are binary (0 or 1),
indicating whether an individual was captured (1) or not (0) for each
capture event.}
\usage{
sample_captures(n_individuals, p_captures)
}
\arguments{
\item{n_individuals}{int: The number of individuals to simulate.}

\item{p_captures}{numeric vector: A vector of capture probabilities for each
capture event. The length of this vector determines the number of capture
events.}
}
\value{
A data frame where each row represents an individual and each column
represents a capture event. The values are binary (0 or 1), indicating
whether an individual was captured (1) or not (0) for each capture event.
}
\description{
Sample Captures for Individuals. This function generates a data frame of
simulated capture data for a specified number of individuals and capture
probabilities. Each row represents an individual, and each column represents
a capture event. The values in the data frame are binary (0 or 1),
indicating whether an individual was captured (1) or not (0) for each
capture event.
}
