% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_control.R
\name{sample_single_capture}
\alias{sample_single_capture}
\title{Sample Captures for a Single Individual. This function generates a binary
vector indicating whether an individual was captured (1) or not (0) for each
capture event based on the provided capture probabilities.}
\usage{
sample_single_capture(capture_probs)
}
\arguments{
\item{capture_probs}{numeric vector: A vector of capture probabilities for
each capture event. The length of this vector determines the number of
capture events.}
}
\value{
A binary vector indicating whether the individual was captured (1)
or not (0) for each capture event.
}
\description{
Sample Captures for a Single Individual. This function generates a binary
vector indicating whether an individual was captured (1) or not (0) for each
capture event based on the provided capture probabilities.
}
\keyword{internal}
