% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_probability.R
\name{summarize_pair}
\alias{summarize_pair}
\title{Summarize the results for a specific pair of capture lists across all
cross-validation folds.}
\usage{
summarize_pair(listpair, folds, diagnostics = "quiet", ...)
}
\arguments{
\item{listpair}{A character string representing the pair of capture lists
(e.g., "1,2").}

\item{folds}{A list of cross-validation folds, each containing training and
test sets.}

\item{diagnostics}{Controls GLM fit diagnostics.}

\item{...}{Additional arguments passed to the `run_fold` function.}
}
\value{
A tibble summarizing the capture probability and variance for the
specified pair of capture lists.
}
\description{
Summarize the results for a specific pair of capture lists across all
cross-validation folds.
}
\keyword{internal}
