% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_control.R
\name{suppress_data}
\alias{suppress_data}
\title{Apply Suppression to Frequency Table Data.
This function takes a frequency table dataset and applies suppression to the
frequency column.}
\usage{
suppress_data(data, suppression_threshold, freq_col = "N_ID")
}
\arguments{
\item{data}{data.frame: the frequency table dataset to apply suppression to}

\item{suppression_threshold}{int: the threshold at which to apply
suppression.}

\item{freq_col}{string: the name of the frequency column in the dataset.
Default is "N_ID".}
}
\value{
a data.frame with suppression applied to the frequency column
}
\description{
Apply Suppression to Frequency Table Data.
This function takes a frequency table dataset and applies suppression to the
frequency column.
}
