% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nuisance.R
\name{tmle_nuisance}
\alias{tmle_nuisance}
\title{Fit TMLE nuisance parameters. This updates the initial estimates using the
TMLE procedure.}
\usage{
tmle_nuisance(q_1, q_2, q_12, y_j, y_k, y_jk, iterations, margin, n_lists)
}
\arguments{
\item{q_1}{Initial estimate for the first nuisance parameter.}

\item{q_2}{Initial estimate for the second nuisance parameter.}

\item{q_12}{Initial estimate for the joint nuisance parameter.}

\item{y_j}{Observed outcome for the first variable.}

\item{y_k}{Observed outcome for the second variable.}

\item{y_jk}{Observed joint outcome for the first and second variables.}

\item{iterations}{Maximum number of iterations for the TMLE update.}

\item{margin}{Margin for estimation.}

\item{n_lists}{Number of lists.}
}
\value{
Updated nuisance parameter estimates.
}
\description{
Fit TMLE nuisance parameters. This updates the initial estimates using the
TMLE procedure.
}
\keyword{internal}
