% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_binary_cols}
\alias{validate_binary_cols}
\title{Validate the capture columns contain only 0 or 1 values}
\usage{
validate_binary_cols(data, end, start = 1)
}
\arguments{
\item{data}{The dataset containing capture columns.}

\item{end}{The number of capture columns to validate.}

\item{start}{The starting column index for the capture columns.}
}
\value{
NULL if all capture columns contain only 0 or 1 values; otherwise,
an error is thrown.
}
\description{
Validate the capture columns contain only 0 or 1 values
}
\keyword{internal}
