% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zero_matrix}
\alias{zero_matrix}
\title{Create a zero matrix with specified dimensions and optional row and column
names.}
\usage{
zero_matrix(nrow, cols, rownames = NULL)
}
\arguments{
\item{nrow}{The number of rows in the matrix.}

\item{cols}{The column names for the matrix.}

\item{rownames}{The row names for the matrix (optional).}
}
\value{
A matrix filled with zeros with the specified dimensions and names.
}
\description{
Create a zero matrix with specified dimensions and optional row and column
names.
}
