% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Function to run the analysis}
\usage{
analyze(n, correlate, suppress, groups, seed = 2024)
}
\arguments{
\item{n}{int: total number entering the simulation}

\item{correlate}{bool: attempts to correlate captures (default is analagous to MA PHD data)}

\item{suppress}{bool: mimic required data suppression (N_ID in 1:10 becomes obfuscated)}

\item{groups}{string: by-group to estimate}

\item{seed}{int: data creation seed}
}
\description{
Function to run the analysis
}
