% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Function to run the analysis}
\usage{
analyze(
  n,
  p_captures,
  p_strata,
  suppress,
  method = "poisson",
  formula.selection = "stepwise",
  opts.stepwise = list(direction = "forward", threshold = 0.5, verbose = FALSE)
)
}
\arguments{
\item{n}{int: total number entering the simulation}

\item{p_captures}{list: list of capture probabilities to emulate}

\item{p_strata}{list: list of extra 'stratification' probabilities to emulate}

\item{suppress}{int: supression cap, defaults to 10}

\item{method}{string: selection for spatial capture recapture model - either "poisson" or "negbin"}

\item{formula.selection}{string: selection for formula decision - either "aic", "corr", or "stepwise"}

\item{opts.stepwise}{list: list containing 'direction' of 'forward' 'backward' or 'both', and 'threshold': p value threshold for stepwise selection}
}
\description{
Function to run the analysis
}
\keyword{internal}
