% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRCSim-package.R
\docType{package}
\name{CRCSim-package}
\alias{CRCSim}
\alias{CRCSim-package}
\title{CRCSim: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
