% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create.data}
\alias{create.data}
\title{Data generation tool}
\usage{
create.data(n_individuals, n_captures, n_strata, p_captures, p_strata)
}
\arguments{
\item{n}{int: number of participants to simulate}

\item{correlate}{bool: attempts correlation between 'captures'}

\item{collapse}{bool: collapse on groups given}

\item{groups}{list: strings of additional cols to group on, supports "agegrp", "sex", and "race"}
}
\value{
a data.table
}
\description{
Data generation tool
}
