% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create.data}
\alias{create.data}
\title{Data generation tool}
\usage{
create.data(n, correlate = FALSE, collapse = TRUE, suppress = FALSE, groups)
}
\arguments{
\item{n}{int: number of participants to simulate}

\item{correlate}{bool: attempts correlation between 'captures'}

\item{collapse}{bool: collapse on groups given}

\item{suppress}{bool: mimic suppression (recode cells in 1:10 as -1)}

\item{groups}{list: strings of additional cols to group on, supports "agegrp", "sex", and "race"}
}
\value{
a data.table
}
\description{
Data generation tool
}
