% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nuisance.R
\name{nuisance_estimation}
\alias{nuisance_estimation}
\title{Estimate nuisance parameters using the specified function.}
\usage{
nuisance_estimation(func, train, test, n_lists, j, k, margin)
}
\arguments{
\item{func}{The function to use for nuisance estimation.}

\item{train}{The training dataset.}

\item{test}{The testing dataset.}

\item{n_lists}{The number of lists.}

\item{j}{Index for the first variable.}

\item{k}{Index for the second variable.}

\item{margin}{The margin for estimation.}
}
\value{
Estimated nuisance parameters.
}
\description{
Estimate nuisance parameters using the specified function.
}
\keyword{internal}
