% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{Options}
\alias{Options}
\title{CRC Options Classes}
\description{
This file defines the R6 classes for the options used in the
different CRC formula selection methods. These classes encapsulate the
parameters and settings for each method, allowing for organized and
structured handling of the options when running CRC with different formula
selection approaches.
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{model}}{Character scalar identifying either the log-linear
model or the nuisance function to utilize.}

    \item{\code{capture_columns}}{Character vector naming the binary capture
indicator columns.}

    \item{\code{threshold}}{Numeric scalar giving the threshold applied by
threshold-based selection methods or the margin desired by stepwise
selection.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Options-initialize}{\code{Options$new()}}
    \item \href{#method-Options-clone}{\code{Options$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Options-initialize}{}}}
\subsection{\code{Options$new()}}{
  Create a new \code{Options} instance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Options$new(model, capture_columns, threshold)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{model}}{Character scalar identifying either the log-linear
model or the nuisance function to utilize.}
      \item{\code{capture_columns}}{Character vector naming the binary capture
indicator columns.}
      \item{\code{threshold}}{Numeric scalar giving the threshold applied by
threshold-based selection methods or the margin desired by stepwise
selection.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The initialized \code{Options} object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Options-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Options-clone}{}}}
\subsection{\code{Options$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Options$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
