% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nuisance.R
\name{qhat_logit}
\alias{qhat_logit}
\title{Estimate the initial nuisance parameters using GLM fits on the training data
and predict on the test data.}
\usage{
qhat_logit(train, test, n_lists, j, k, margin)
}
\arguments{
\item{train}{The training dataset.}

\item{test}{The testing dataset.}

\item{n_lists}{The number of lists.}

\item{j}{Index for the first variable.}

\item{k}{Index for the second variable.}

\item{margin}{The margin for estimation.}
}
\value{
A list containing the initial estimates for q_1, q_2, and q_12.
}
\description{
Estimate the initial nuisance parameters using GLM fits on the training data
and predict on the test data.
}
\keyword{internal}
