% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossfold.R
\name{run_fold}
\alias{run_fold}
\title{Run a single fold of the cross-validation procedure.}
\usage{
run_fold(
  fold_split,
  method,
  j,
  k,
  margin,
  n_lists,
  n,
  nuisance_estimation_func
)
}
\arguments{
\item{fold_split}{A list containing the training set, test set, and the
subset of indices for the current fold.}

\item{method}{The estimation method to be used.}

\item{j}{The index of the first list.}

\item{k}{The index of the second list.}

\item{margin}{The margin parameter for estimation.}

\item{n_lists}{The total number of capture lists.}

\item{n}{The total number of observations.}

\item{nuisance_estimation_func}{The function used for nuisance parameter
estimation.}

\item{...}{Additional arguments passed to the nuisance estimation function.}
}
\value{
A list containing the capture probability and its variance for the
current fold.
}
\description{
Run a single fold of the cross-validation procedure.
}
