% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCRCsim.R
\name{runCRC}
\alias{runCRC}
\title{Function to run CRC sim - returns a list of data.tables resulting from \code{analyze}}
\usage{
runCRC(nboot, ncores, p_captures, p_strata, seed = 2024)
}
\arguments{
\item{nboot}{int: number of bootstraps}

\item{ncores}{int: number of cores to use}

\item{p_captures}{list: list of capture probabilities}

\item{p_strata}{list: list of strata probabilities - for nonstratified just use '1'}

\item{seed}{int: starting seed}
}
\description{
Function to run CRC sim - returns a list of data.tables resulting from \code{analyze}
}
