% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Run the CRC Simulations for the provided number of individuals, covariates,
and captures.}
\usage{
simulate(
  n_individuals,
  n_covariates,
  captures,
  cov_func = covariate_conditions
)
}
\arguments{
\item{n_individuals}{The number of individuals to simulate.}

\item{n_covariates}{The covariates that accompany the captures.}

\item{captures}{the captures that accompany each covariate.}
}
\value{
The results of the CRC Simulation
}
\description{
Run the CRC Simulations for the provided number of individuals, covariates,
and captures.
}
