% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Attempts recovery of the ground truth - fundamentally a wrapper for Syndemics::crc}
\usage{
simulate(
  DT,
  capture = c("APCD", "BSAS", "Casemix", "Death", "Matris", "PMP"),
  group,
  suppress,
  method,
  formula.selection,
  opts.stepwise
)
}
\arguments{
\item{DT}{data.table from the \code{create.data} step}

\item{capture}{list: strings of captures}

\item{group}{stratification string to group by}

\item{suppress}{numeric: maximum value to suppress at}

\item{method}{string: selection for spatial capture recapture model - either "poisson" or "negbin"}

\item{formula.selection}{string: selection for formula decision - either "aic", "corr", or "stepwise"}

\item{opts.stepwise}{list: list containing 'direction' of 'forward' 'backward' or 'both', and 'threshold': p value threshold for stepwise selection}
}
\description{
Attempts recovery of the ground truth - fundamentally a wrapper for Syndemics::crc
}
\keyword{internal}
