% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Attempts recovery of the ground truth - fundamentally a wrapper for Syndemics::crc}
\usage{
simulate(
  DT,
  capture = c("APCD", "BSAS", "Casemix", "Death", "Matris", "PMP"),
  groups,
  ...
)
}
\arguments{
\item{DT}{data.table from the \code{create.data} step}

\item{capture}{list: strings of captures}

\item{groups}{stratification string to group by}

\item{...}{other options, as passed to Syndemics::crc}
}
\description{
Attempts recovery of the ground truth - fundamentally a wrapper for Syndemics::crc
}
