% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_list_overlap}
\alias{test_list_overlap}
\title{Test the overlap between two capture lists.}
\usage{
test_list_overlap(l1, l2, margin)
}
\arguments{
\item{l1}{The first capture list.}

\item{l2}{The second capture list.}

\item{margin}{The minimum acceptable overlap between the lists.}
}
\value{
NULL. A warning is issued if the overlap is less than the margin.
}
\description{
Test the overlap between two capture lists.
}
